refactor: Context API로 상태 공유하기
